app/requirements.txt
pandas
fastapi
uvicorn
python-multipart
python-dotenv
langgraph
langchain-openai
openpyxl